refactor: created custom hook for fetching of albums
