+++ src/lib/spotifyApi.ts
@@ -12,32 +12,6 @@ const getAlbum = async (access_token: string, offset: number) => {
   return data;
 };
 
-const getAlbums = async (
-  access_token: string,
-  expires_at: number,
-  { pageParam }: { pageParam: { offset: number; limit: number } },
-) => {
-  if (Date.now() >= expires_at) {
-    throw new Error('Access Token is invalid');
-  }
-
-  const data = await fetch(
-    `https://api.spotify.com/v1/me/albums?offset=${pageParam.offset}&limit=${pageParam.limit}&locale=*`,
-    {
-      method: 'GET',
-      headers: {
-        Authorization: `Bearer ${access_token}`,
-      },
-    },
-  );
-
-  if (!data) {
-    throw new Error();
-  }
-
-  return (await data.json()) as Albums;
-};
-
 const putPlayback = (access_token: string, uri: string) => {
   return fetch('https://api.spotify.com/v1/me/player/play', {
     method: 'PUT',
@@ -50,4 +24,4 @@ const putPlayback = (access_token: string, uri: string) => {
   });
 };
 
-export { getAlbum, getAlbums, putPlayback };
+export { getAlbum, putPlayback };
